+++ server/server/routing/db_routes.js
@@ -45,6 +45,11 @@ const actionsRoutes = require("./routes/actions_routes");
 const projectsRoutes = require("./routes/projects_routes");
 const archivesRoutes = require("./routes/archives_routes");
 const errorLogsRoutes = require("./routes/error_logs_routes");
+const usersRoutes = require("./routes/users_routes");
+const timeLoggingRoutes = require("./routes/time_logging_routes");
+const submissionsRoutes = require("./routes/submissions_routes");
+const proposalsRoutes = require("./routes/proposals_routes");
+const filesRoutes = require("./routes/files_routes");
 
 // Mount split routers early so they take precedence over the large monolithic definitions
 db_router.use("/", sponsorsRoutes(db));
@@ -52,6 +57,11 @@ db_router.use("/", actionsRoutes(db));
 db_router.use("/", projectsRoutes(db));
 db_router.use("/", archivesRoutes(db));
 db_router.use("/", errorLogsRoutes(db));
+db_router.use("/", usersRoutes(db));
+db_router.use("/", timeLoggingRoutes(db));
+db_router.use("/", submissionsRoutes(db));
+db_router.use("/", proposalsRoutes(db));
+db_router.use("/", filesRoutes(db));
 const CONSTANTS = require("../consts");
 const { ROLES } = require("../consts");
 const { off } = require("process");
@@ -97,1124 +107,6 @@ module.exports = (db) => {
       }
     });
   }
-
-  db_router.get(
-    "/selectAllStudentInfo",
-    [UserAuth.isCoachOrAdmin],
-    (req, res, next) => {
-      let getStudentsQuery = `
-            SELECT *
-            FROM users
-            LEFT JOIN semester_group
-            ON users.semester_group = semester_group.semester_id
-            WHERE type = 'student'
-            ORDER BY semester_group desc
-        `;
-      db.query(getStudentsQuery)
-        .then((values) => {
-          res.send(values);
-        })
-        .catch((err) => {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get(
-    "/selectAllNonStudentInfo",
-    [UserAuth.isAdmin],
-    (req, res, next) => {
-      let getUsersQuery = `
-            SELECT *
-            FROM users
-            LEFT JOIN semester_group
-            ON users.semester_group = semester_group.semester_id
-            WHERE type != 'student'
-        `;
-      db.query(getUsersQuery)
-        .then((values) => {
-          res.send(values);
-        })
-        .catch((err) => {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get(
-    "/getSemesterStudents",
-    [UserAuth.isSignedIn],
-    (req, res, next) => {
-      let query = "";
-      let params = [];
-      switch (req.user.type) {
-        //Retrieves all users from a semester group that is similar to the student that is making the query.
-        case ROLES.STUDENT:
-          query = `
-            SELECT users.* 
-            FROM users 
-            WHERE users.semester_group = (
-              SELECT semester_group FROM users WHERE system_id = ?
-            ) AND users.type = 'student'`;
-          params = [req.user.system_id];
-          break;
-
-        case ROLES.COACH:
-          query = `
-            SELECT users.* FROM users
-            LEFT JOIN semester_group
-              ON users.semester_group = semester_group.semester_id
-            WHERE users.semester_group IN (
-              SELECT projects.semester FROM projects
-              WHERE projects.project_id IN (
-                SELECT project_coaches.project_id FROM project_coaches
-                WHERE project_coaches.coach_id = ?
-              )
-            )`;
-          params = [req.user.system_id];
-          break;
-
-        case ROLES.ADMIN:
-          query = `SELECT * FROM users
-                    LEFT JOIN semester_group
-                    ON users.semester_group = semester_group.semester_id
-                    WHERE users.type = 'student'`;
-          break;
-        default:
-          break;
-      }
-
-      db.query(query, params)
-        .then((users) => {
-          if (req.user.type === ROLES.STUDENT) {
-            users = users.map((user) => {
-              let output = {};
-              if (user.project === req.user.project) {
-                output["last_login"] = user["last_login"];
-                output["prev_login"] = user["prev_login"];
-              }
-              output["active"] = user["active"];
-              output["email"] = user["email"];
-              output["fname"] = user["fname"];
-              output["lname"] = user["lname"];
-              output["project"] = user["project"];
-              output["semester_group"] = user["semester_group"];
-              output["system_id"] = user["system_id"];
-              output["type"] = user["type"];
-              return output;
-            });
-          }
-          res.send(users);
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get("/getProjectMembers", [UserAuth.isSignedIn], (req, res) => {
-    let query = `SELECT users.*, project_coaches.project_id FROM users
-            LEFT JOIN project_coaches ON project_coaches.coach_id = users.system_id
-            WHERE users.project = ? OR project_coaches.project_id = ?`;
-
-    params = [req.query.project_id, req.query.project_id];
-
-    db.query(query, params).then((users) => res.send(users));
-  });
-
-  // NOTE: This is currently used for getting user for AdminView to mock users, however, I feel that this network request will get quite large
-  // as we add about 100 users every semester.
-  db_router.get("/getActiveUsers", [UserAuth.isAdmin], (req, res) => {
-    let query = `SELECT ${CONSTANTS.SIGN_IN_SELECT_ATTRIBUTES}
-            FROM users
-            WHERE active = ''`;
-    db.query(query).then((users) => res.send(users));
-  });
-
-  db_router.post(
-    "/createUser",
-    [
-      UserAuth.isAdmin,
-      UserAuth.canWrite,
-      body("system_id")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 50 }),
-      body("fname")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 50 }),
-      body("lname")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 50 }),
-      body("email")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 50 }),
-      body("type")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 50 }),
-      body("semester_group").isLength({ max: 50 }),
-      body("project").isLength({ max: 50 }),
-      body("active").trim().escape().isLength({ max: 50 }),
-      body("viewOnly").trim().escape().isLength({ max: 50 }),
-    ],
-    async (req, res, next) => {
-      let result = validationResult(req);
-      console.log(result);
-
-      if (result.errors.length !== 0) {
-        const error = new Error("Validation Error");
-        error.statusCode = 400;
-        return next(error);
-      }
-
-      let body = req.body;
-
-      const sql = `INSERT INTO ${DB_CONFIG.tableNames.users}
-                (system_id, fname, lname, email, type, semester_group, project, active, view_only, profile_info)
-                VALUES (?,?,?,?,?,?,?,?,?,?)`;
-
-      const active =
-        body.active === "false"
-          ? moment().format(CONSTANTS.datetime_format)
-          : "";
-
-      const viewOnly = body.viewOnly === "true" ? "TRUE" : "FALSE";
-
-      // Default profile_info with required fields
-      const defaultProfileInfo = JSON.stringify({
-        additional_info: "",
-        dark_mode: false,
-        gantt_view: true,
-      });
-
-      const params = [
-        body.system_id,
-        body.fname,
-        body.lname,
-        body.email,
-        body.type,
-        body.semester_group === "" ? null : body.semester_group,
-        body.project === "" ? null : body.project,
-        active,
-        viewOnly,
-        defaultProfileInfo,
-      ];
-      db.query(sql, params)
-        .then(() => {
-          return res.status(200).send();
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.post(
-    "/batchCreateUser",
-    [
-      UserAuth.isAdmin,
-      UserAuth.canWrite,
-      // TODO: Add more validation
-    ],
-    async (req, res, next) => {
-      try {
-        let users = JSON.parse(req.body.users);
-        const failedUsers = [];
-        const successUsers = [];
-
-        for (const user of users) {
-          // Default profile_info with required fields
-          const defaultProfileInfo = JSON.stringify({
-            additional_info: "",
-            dark_mode: false,
-            gantt_view: true,
-          });
-
-          const values = [
-            user.system_id,
-            user.fname,
-            user.lname,
-            user.email,
-            user.type,
-            user.semester_group === "" ? null : user.semester_group,
-            user.active.toLocaleLowerCase() === "false"
-              ? moment().format(CONSTANTS.datetime_format)
-              : "",
-            defaultProfileInfo,
-          ];
-
-          try {
-            await db.query(
-              `INSERT INTO ${DB_CONFIG.tableNames.users} 
-              (system_id, fname, lname, email, type, semester_group, active, profile_info) 
-              VALUES (?, ?, ?, ?, ?, ?, ?, ?)`,
-              values,
-            );
-            successUsers.push(user);
-          } catch (err) {
-            let errorMessage = err.message;
-
-            // Provide more user-friendly error messages for common constraint violations
-            if (err.code === "SQLITE_CONSTRAINT") {
-              if (
-                err.message.includes(
-                  "UNIQUE constraint failed: users.system_id",
-                )
-              ) {
-                errorMessage = `System ID '${user.system_id}' already exists`;
-              } else if (
-                err.message.includes("UNIQUE constraint failed: users.email")
-              ) {
-                errorMessage = `Email '${user.email}' already exists`;
-              } else {
-                errorMessage = "Duplicate data - user may already exist";
-              }
-            }
-
-            failedUsers.push({ user, error: errorMessage });
-          }
-        }
-
-        res.status(200).json({ successUsers, failedUsers });
-      } catch (err) {
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      }
-    },
-  );
-
-  db_router.post(
-    "/editUser",
-    [UserAuth.isAdmin, UserAuth.canWrite],
-    (req, res, next) => {
-      let body = req.body;
-
-      let updateQuery = `
-            UPDATE users
-            SET fname = ?,
-                lname = ?,
-                email = ?,
-                type = ?,
-                semester_group = ?,
-                project = ?,
-                active = ?,
-                view_only = ?
-            WHERE system_id = ?
-        `;
-
-      const active =
-        body.active === "false"
-          ? moment().format(CONSTANTS.datetime_format)
-          : "";
-
-      const viewOnly = body.viewOnly === "true" ? "TRUE" : "FALSE";
-
-      let params = [
-        body.fname,
-        body.lname,
-        body.email,
-        body.type,
-        body.semester_group || null,
-        body.project || null,
-        active,
-        viewOnly,
-        body.system_id,
-      ];
-
-      db.query(updateQuery, params)
-        .then(() => {
-          return res.status(200).send();
-        })
-        .catch((err) => {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.post(
-    "/removeTime",
-    [UserAuth.isSignedIn, UserAuth.canWrite],
-    (req, res, next) => {
-      if (!req.body.id) {
-        const error = new Error("No Id Provided");
-        error.statusCode = 400;
-        return next(error);
-      }
-
-      const sql = "UPDATE time_log SET active=0 WHERE time_log_id = ?";
-
-      db.query(sql, [req.body.id])
-        .then(() => {
-          res.status(200).send();
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get("/avgTime", [UserAuth.isSignedIn], async (req, res, next) => {
-    const sql =
-      "SELECT ROUND(AVG(CASE WHEN active != 0 THEN time_amount ELSE NULL END), 2)  AS avgTime, system_id FROM time_log WHERE project = ? GROUP BY system_id";
-    console.log(req.query.project_id);
-
-    db.query(sql, [req.query.project_id])
-      .then((time) => {
-        console.log(time);
-        res.send(time);
-      })
-      .catch((err) => {
-        console.error(err);
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      });
-  });
-
-  db_router.post(
-    "/createTimeLog",
-    [UserAuth.canWrite],
-    async (req, res, next) => {
-      let result = validationResult(req);
-
-      if (result.errors.length !== 0) {
-        const error = new Error("Validation Error");
-        error.statusCode = 400;
-        return next(error);
-      }
-
-      // Validate that the work date is not in the future
-      // This prevents users from logging time for dates that haven't occurred yet
-      const workDate = new Date(req.body.date);
-      const currentDate = new Date();
-      const currentDateOnly = new Date(
-        currentDate.getFullYear(),
-        currentDate.getMonth(),
-        currentDate.getDate(),
-      );
-      const workDateOnly = new Date(
-        workDate.getFullYear(),
-        workDate.getMonth(),
-        workDate.getDate(),
-      );
-
-      if (workDateOnly > currentDateOnly) {
-        const error = new Error("Cannot log time for future dates");
-        error.statusCode = 400;
-        return next(error);
-      }
-
-      // Validate that the work date is within the past 14 days (2 weeks)
-      // This maintains the existing business rule about recent time logging
-      const twoWeeksAgo = new Date(currentDateOnly);
-      twoWeeksAgo.setDate(twoWeeksAgo.getDate() - 14);
-
-      if (workDateOnly < twoWeeksAgo) {
-        const error = new Error("Cannot log time for dates older than 14 days");
-        error.statusCode = 400;
-        return next(error);
-      }
-
-      let mock_id = req.user.mock ? req.user.mock.system_id : "";
-
-      const sql = `INSERT INTO time_log
-                (semester, system_id, project, mock_id, work_date, time_amount, work_comment)
-                VALUES (?,?,?,?,?,?,?)`;
-
-      const params = [
-        req.user.semester_group,
-        req.user.system_id,
-        req.user.project,
-        mock_id,
-        req.body.date,
-        req.body.time_amount,
-        req.body.comment,
-      ];
-      db.query(sql, params)
-        .then(() => {
-          return res.status(200).send();
-        })
-        .catch((err) => {
-          console.error(err);
-          let error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  /**
-   * Responds with a list of links to pdf versions of proposal forms
-   *
-   * NOTE: This route is unused and untested.
-   */
-  db_router.get(
-    "/getProposalPdfNames",
-    UserAuth.isSignedIn,
-    (req, res, next) => {
-      fs.readdir(
-        path.join(__dirname, "../proposal_docs"),
-        function (err, files) {
-          if (err) {
-            const error = new Error(err);
-            error.statusCode = 500;
-            return next(error);
-          }
-          let fileLinks = [];
-          files.forEach(function (file) {
-            fileLinks.push(file.toString());
-          });
-
-          res.send(fileLinks);
-        },
-      );
-    },
-  );
-
-  db_router.get("/getProposalPdf", UserAuth.isSignedIn, (req, res) => {
-    if (req.query.project_id) {
-      let projectId = req.query.project_id.replace(/\\|\//g, ""); // attempt to avoid any path traversal issues
-      res.sendFile(path.join(__dirname, `../proposal_docs/${projectId}.pdf`));
-    } else res.send("File not found");
-  });
-
-  // NOTE: This route is unused and untested.
-  db_router.get(
-    "/getProposalAttachmentNames",
-    UserAuth.isSignedIn,
-    (req, res, next) => {
-      if (req.query.project_id) {
-        let projectId = req.query.project_id.replace(/\\|\//g, ""); // attempt to avoid any path traversal issues, get the name with no extension
-        fs.readdir(
-          path.join(__dirname, `./server/sponsor_proposal_files/${projectId}`),
-          function (err, files) {
-            if (err) {
-              const error = new Error(err);
-              error.statusCode = 500;
-              return next(error);
-            }
-            let fileLinks = [];
-            files.forEach(function (file) {
-              fileLinks.push(file.toString());
-            });
-
-            res.send(fileLinks);
-          },
-        );
-      } else {
-        res.status(404).send("Bad request");
-      }
-    },
-  );
-
-  db_router.get("/getProposalAttachment", UserAuth.isSignedIn, (req, res) => {
-    if (req.query.project_id && req.query.name) {
-      let projectId = req.query.project_id.replace(/\\|\//g, ""); // attempt to avoid any path traversal issues
-      let name = req.query.name.replace(/\\|\//g, ""); // attempt to avoid any path traversal issues
-      res.sendFile(
-        path.join(__dirname, `../sponsor_proposal_files/${projectId}/${name}`),
-      );
-    } else res.send("File not found");
-  });
-
-  /**
-   * WARN: THIS IS VERY DANGEROUS AND IT CAN BE USED TO OVERWRITE SERVER FILES.
-   */
-  db_router.post(
-    "/uploadFiles",
-    [UserAuth.isAdmin, UserAuth.canWrite],
-    (req, res, next) => {
-      let filesUploaded = [];
-
-      // Attachment Handling
-      if (req.files && req.files.files) {
-        // If there is only one attachment, then it does not come as a list
-        if (req.files.files.length === undefined) {
-          req.files.files = [req.files.files];
-        }
-
-        const formattedPath = `resource/${req.body.path}`;
-        const baseURL = path.join(__dirname, `../../${formattedPath}`);
-
-        //If directory, exists, it won't make one, otherwise it will based on the baseUrl :/
-        fs.mkdirSync(baseURL, { recursive: true });
-        for (let x = 0; x < req.files.files.length; x++) {
-          req.files.files[x].mv(
-            `${baseURL}/${req.files.files[x].name}`,
-            function (err) {
-              if (err) {
-                console.error(err);
-                const error = new Error(err);
-                error.statusCode = 500;
-                return next(error);
-              }
-            },
-          );
-          filesUploaded.push(
-            `${process.env.BASE_URL}/${formattedPath}/${req.files.files[x].name}`,
-          );
-        }
-      }
-      res.send({ msg: "Success!", filesUploaded: filesUploaded });
-    },
-  );
-
-  /**
-   * WARN: THIS IS VERY DANGEROUS AND IT CAN BE USED TO OVERWRITE SERVER FILES.
-   */
-  db_router.post("/uploadFilesStudent", UserAuth.canWrite, (req, res, next) => {
-    let filesUploaded = [];
-
-    // Attachment Handling
-    if (req.files && req.files.files) {
-      // If there is only one attachment, then it does not come as a list
-      if (req.files.files.length === undefined) {
-        req.files.files = [req.files.files];
-      }
-
-      const formattedPath = `resource/${req.body.path}`;
-      const baseURL = path.join(__dirname, `../../${formattedPath}`);
-
-      //If directory, exists, it won't make one, otherwise it will based on the baseUrl :/
-      fs.mkdirSync(baseURL, { recursive: true });
-      for (let x = 0; x < req.files.files.length; x++) {
-        req.files.files[x].mv(
-          `${baseURL}/${req.files.files[x].name}`,
-          function (err) {
-            if (err) {
-              const error = new Error(err);
-              error.statusCode = 500;
-              return next(error);
-            }
-          },
-        );
-        filesUploaded.push(
-          `${process.env.BASE_URL}/${formattedPath}/${req.files.files[x].name}`,
-        );
-        let fileName = req.files.files[x].name;
-        let pathString = req.body.path;
-        pathString = pathString.split("/");
-        pathString.shift();
-        pathString = '"' + pathString.join("/") + "/" + fileName + '"';
-        let query = `UPDATE ${DB_CONFIG.tableNames.archive}
-                     SET ${req.body.column} = ${pathString}
-                     WHERE archive_id = ${req.body.archive}`;
-        db.query(query).catch((err) => {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-      }
-    }
-    res.send({ msg: "Success!", filesUploaded: filesUploaded });
-  });
-
-  db_router.post(
-    "/createDirectory",
-    [UserAuth.isAdmin, UserAuth.canWrite],
-    (req, res, next) => {
-      const formattedPath =
-        req.query.path === "" ? `resource/` : `resource/${req.query.path}`;
-      const baseURL = path.join(__dirname, `../../${formattedPath}`);
-      if (!fs.existsSync(baseURL)) {
-        fs.mkdirSync(baseURL, { recursive: true });
-        res.send({ msg: "Success!" });
-      } else {
-        const error = new Error("Directory already exists");
-        error.statusCode = 500;
-        return next(error);
-      }
-    },
-  );
-
-  db_router.post(
-    "/renameDirectoryOrFile",
-    [UserAuth.isAdmin, UserAuth.canWrite],
-    (req, res, next) => {
-      const { oldPath, newPath } = req.query;
-      const formattedOldPath =
-        oldPath === "" ? `resource/` : `resource/${oldPath}`;
-      const formattedNewPath =
-        newPath === "" ? `resource/` : `resource/${newPath}`;
-      const baseURLOld = path.join(__dirname, `../../${formattedOldPath}`);
-      const baseURLNew = path.join(__dirname, `../../${formattedNewPath}`);
-
-      // New path already exists, so we can't rename
-      if (fs.existsSync(baseURLNew)) {
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      }
-      // Copy all files from old directory to new directory
-      if (fs.lstatSync(baseURLOld).isDirectory()) {
-        fse.copySync(baseURLOld, baseURLNew);
-        fs.rmdirSync(baseURLOld, { recursive: true });
-        res.send({ msg: "Success!" });
-        // Rename file
-      } else if (fs.lstatSync(baseURLOld).isFile()) {
-        fs.renameSync(baseURLOld, baseURLNew);
-        res.send({ msg: "Success!" });
-      }
-    },
-  );
-
-  db_router.get("/getFiles", UserAuth.isAdmin, (req, res, next) => {
-    let fileData = [];
-    // This is the path with the specified directory we want to find files in.
-    const formattedPath =
-      req.query.path === "" ? `resource/` : `resource/${req.query.path}`;
-    const baseURL = path.join(__dirname, `../../${formattedPath}`);
-
-    if (fs.existsSync(baseURL)) {
-      // Get the files in the directory
-      fs.readdir(baseURL, function (err, files) {
-        if (err) {
-          console.error(`Error reading directory ${baseURL}:`, err);
-          // Return empty array instead of throwing error for missing directories
-          res.send(fileData);
-          return;
-        }
-
-        try {
-          const info = fs.statSync(baseURL);
-          files.forEach(function (file) {
-            try {
-              // Only files have sizes, directories do not. Send file size if it is a file
-              const fileInfo = fs.statSync(path.join(baseURL, file));
-              if (fileInfo.isFile()) {
-                fileData.push({
-                  file: file,
-                  size: fileInfo.size,
-                  lastModified: fileInfo.ctime,
-                });
-              } else {
-                fileData.push({
-                  file: file,
-                  size: 0,
-                  lastModified: info.ctime,
-                });
-              }
-            } catch (fileErr) {
-              console.error(`Error processing file ${file}:`, fileErr);
-              // Skip files that can't be processed
-            }
-          });
-        } catch (statErr) {
-          console.error(
-            `Error getting directory stats for ${baseURL}:`,
-            statErr,
-          );
-        }
-
-        res.send(fileData);
-      });
-    } else {
-      console.log(`Directory does not exist: ${baseURL}`);
-      res.send(fileData);
-    }
-  });
-
-  db_router.get("/getProjectFiles", (req, res, next) => {
-    let fileData = [];
-    // This is the path with the specified directory we want to find files in.
-    const formattedPath = `resource/`;
-    const baseURL = path.join(__dirname, `../../${formattedPath}`);
-    fs.mkdirSync(baseURL, { recursive: true });
-    // Get the files in the directory
-    fs.readdir(baseURL, function (err, files) {
-      if (err) {
-        console.error(err);
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      }
-      const info = fs.statSync(baseURL);
-      files.forEach(function (file) {
-        // Only files have sizes, directories do not. Send file size if it is a file
-        const fileInfo = fs.statSync(baseURL + file);
-        if (fileInfo.isFile()) {
-          fileData.push({
-            file: file,
-            size: fileInfo.size,
-            lastModified: fileInfo.ctime,
-          });
-        } else {
-          fileData.push({
-            file: file,
-            size: 0,
-            lastModified: info.ctime,
-          });
-        }
-      });
-      res.send(fileData);
-    });
-  });
-
-  db_router.get("/getProjectFiles", (req, res, next) => {
-    let fileData = [];
-    // This is the path with the specified directory we want to find files in.
-    const formattedPath = `resource/`;
-    const baseURL = path.join(__dirname, `../../${formattedPath}`);
-    fs.mkdirSync(baseURL, { recursive: true });
-    // Get the files in the directory
-    fs.readdir(baseURL, function (err, files) {
-      if (err) {
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      }
-      const info = fs.statSync(baseURL);
-      files.forEach(function (file) {
-        // Only files have sizes, directories do not. Send file size if it is a file
-        const fileInfo = fs.statSync(baseURL + file);
-        if (fileInfo.isFile()) {
-          fileData.push({
-            file: file,
-            size: fileInfo.size,
-            lastModified: fileInfo.ctime,
-          });
-        } else {
-          fileData.push({
-            file: file,
-            size: 0,
-            lastModified: info.ctime,
-          });
-        }
-      });
-      res.send(fileData);
-    });
-  });
-
-  db_router.delete(
-    "/removeFile",
-    [UserAuth.isAdmin, UserAuth.canWrite],
-    (req, res, next) => {
-      const formattedPath = `resource/${req.query.path}`;
-      const baseURL = path.join(__dirname, `../../${formattedPath}`);
-      fs.unlink(baseURL, (err) => {
-        if (err) {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        } else {
-          res.send({ msg: "Success!" });
-        }
-      });
-    },
-  );
-
-  db_router.delete(
-    "/removeDirectory",
-    [UserAuth.isAdmin, UserAuth.canWrite],
-    (req, res, next) => {
-      const formattedPath = `resource/${req.query.path}`;
-      const baseURL = path.join(__dirname, `../../${formattedPath}`);
-      if (fs.existsSync(baseURL)) {
-        fs.rmdirSync(baseURL, { recursive: true });
-        return res.status(200).send({ msg: "Success!" });
-      } else {
-        const error = new Error("Directory does not exist, cannot delete");
-        error.statusCode = 500;
-        return next(error);
-      }
-    },
-  );
-
-  db_router.post(
-    "/submitProposal",
-    [
-      // TODO: Should the max length be set to something smaller than 5000?
-      body("title")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 50 })
-        .withMessage("Title must be under 50 characters"),
-      body("organization")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("primary_contact")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("contact_email")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("contact_phone")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("background_info")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("project_description")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("project_scope")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("project_challenges")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("sponsor_provided_resources")
-        .trim()
-        .escape()
-        .isLength({ max: 5000 }),
-      body("constraints_assumptions")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("sponsor_deliverables")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("proprietary_info").trim().escape().isLength({ max: 5000 }),
-      body("sponsor_alternate_time").trim().escape().isLength({ max: 5000 }),
-      body("sponsor_avail_checked")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      body("project_agreements_checked")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      body("assignment_of_rights")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-    ],
-    async (req, res, next) => {
-      let result = validationResult(req);
-
-      if (result.errors.length !== 0) {
-        const errorMessages = result.errors
-          .map((error) => `${error.param}: ${error.msg}`)
-          .join(", ");
-        const error = new Error(`Validation failed: ${errorMessages}`);
-        error.statusCode = 400;
-        return next(error);
-      }
-
-      // Insert into the database
-      let body = req.body;
-
-      let date = new Date();
-      let timeString = `${date.getFullYear()}-${date.getUTCMonth()}-${date.getDate()}`;
-      const projectId = `${timeString}_${nanoid()}`;
-
-      let filenamesCSV = "";
-      // Attachment Handling
-      if (req.files && req.files.attachments) {
-        // If there is only one attachment, then it does not come as a list
-        if (req.files.attachments.length === undefined) {
-          req.files.attachments = [req.files.attachments];
-        }
-
-        if (req.files.attachments.length > 5) {
-          // Don't allow more than 5 files
-          const error = new Error("Maximum of 5 files allowed");
-          error.statusCode = 400;
-          return next(error);
-        }
-
-        const baseURL = path.join(
-          __dirname,
-          `../sponsor_proposal_files/${projectId}`,
-        );
-
-        fs.mkdirSync(baseURL, { recursive: true });
-
-        for (let x = 0; x < req.files.attachments.length; x++) {
-          if (req.files.attachments[x].size > 15 * 1024 * 1024) {
-            // 15mb limit exceeded
-            const error = new Error("File size limit exceeded");
-            error.statusCode = 400;
-            return next(error);
-          }
-          if (
-            !CONFIG.accepted_file_types.includes(
-              path.extname(req.files.attachments[x].name),
-            )
-          ) {
-            // send an error if the file is not an accepted type
-            const error = new Error("file type not accepted");
-            error.statusCode = 400;
-            return next(error);
-          }
-
-          // Append the file name to the CSV string, begin with a comma if x is not 0
-          filenamesCSV +=
-            x === 0
-              ? `${req.files.attachments[x].name}`
-              : `, ${req.files.attachments[x].name}`;
-
-          req.files.attachments[x].mv(
-            `${baseURL}/${req.files.attachments[x].name}`,
-            function (err) {
-              if (err) {
-                console.error(err);
-                const error = new Error(err);
-                error.statusCode = 500;
-                return next(error);
-              }
-            },
-          );
-        }
-      }
-      const sql = `INSERT INTO ${DB_CONFIG.tableNames.senior_projects}
-                (project_id, status, title, organization, primary_contact, contact_email, contact_phone, attachments,
-                background_info, project_description, project_scope, project_challenges,
-                sponsor_provided_resources, constraints_assumptions, sponsor_deliverables,
-                proprietary_info, sponsor_alternate_time, sponsor_avail_checked, project_agreements_checked, assignment_of_rights)
-                VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)`;
-
-      const params = [
-        projectId,
-        "submitted",
-        body.title.substring(0, 50),
-        body.organization,
-        body.primary_contact,
-        body.contact_email,
-        body.contact_phone,
-        filenamesCSV,
-        body.background_info,
-        body.project_description,
-        body.project_scope,
-        body.project_challenges,
-        body.sponsor_provided_resources,
-        body.constraints_assumptions,
-        body.sponsor_deliverables,
-        body.proprietary_info,
-        body.sponsor_alternate_time,
-        body.sponsor_avail_checked,
-        body.project_agreements_checked,
-        body.assignment_of_rights,
-      ];
-
-      db.query(sql, params)
-        .then(() => {
-          let doc = new PDFDoc();
-          const baseURL = path.join(__dirname, `../proposal_docs/`);
-          fs.mkdirSync(baseURL, { recursive: true });
-          doc.pipe(fs.createWriteStream(`${baseURL}/${projectId}.pdf`));
-
-          doc.font("Times-Roman");
-
-          for (let key of Object.keys(DB_CONFIG.senior_project_proposal_keys)) {
-            doc
-              .fill("blue")
-              .fontSize(16)
-              .text(DB_CONFIG.senior_project_proposal_keys[key]),
-              {
-                underline: true,
-              };
-            doc
-              .fontSize(12)
-              .fill("black")
-              .text(convert(he.decode(body[key] || ""))); // Text value from proposal
-            doc.moveDown();
-            doc.save();
-          }
-
-          doc.fill("blue").fontSize(16).text("Attachments"),
-            {
-              underline: true,
-            };
-          doc.fontSize(12).fill("black").text(filenamesCSV);
-          doc.moveDown();
-          doc.save();
-
-          doc.end();
-          return res.status(200).send();
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
 
   db_router.get(
     "/getActiveTimelines",
@@ -1875,115 +767,6 @@ module.exports = (db) => {
           error.statusCode = 500;
           return next(error);
         });
-    },
-  );
-
-  db_router.get("/getSubmission", [UserAuth.isSignedIn], (req, res, next) => {
-    let getSubmissionQuery = "";
-    let params = [];
-
-    switch (req.user.type) {
-      case ROLES.STUDENT:
-        getSubmissionQuery = `SELECT action_log.form_data, action_log.files
-                    FROM action_log
-                    JOIN actions ON actions.action_id = action_log.action_template
-                    WHERE action_log.action_log_id = ? AND (actions.action_target = '${ACTION_TARGETS.TEAM}' OR action_log.system_id = ?)`;
-        params = [req.query.log_id, req.user.system_id];
-        break;
-      case ROLES.COACH:
-        getSubmissionQuery = `SELECT action_log.form_data, action_log.files
-                    FROM action_log
-                    JOIN project_coaches ON project_coaches.project_id = action_log.project
-                    WHERE action_log.action_log_id = ? AND project_coaches.coach_id = ?`;
-        params = [req.query.log_id, req.user.system_id];
-        break;
-      case ROLES.ADMIN:
-        getSubmissionQuery = `SELECT action_log.form_data, action_log.files
-                    FROM action_log
-                    WHERE action_log.action_log_id = ?`;
-        params = [req.query.log_id];
-        break;
-      default:
-        const error = new Error("Unknown Role");
-        error.statusCode = 401;
-        return next(error);
-    }
-
-    db.query(getSubmissionQuery, params)
-      .then((submissions) => {
-        res.send(submissions);
-      })
-      .catch((err) => {
-        console.error(err);
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      });
-  });
-
-  db_router.get(
-    "/getSubmissionFile",
-    [UserAuth.isSignedIn],
-    async (req, res, next) => {
-      let getSubmissionQuery = "";
-      let params = [];
-
-      switch (req.user.type) {
-        case ROLES.STUDENT:
-          getSubmissionQuery = `SELECT action_log.files, action_log.project, action_log.system_id, actions.action_id, actions.action_target
-                    FROM action_log
-                    JOIN actions ON actions.action_id = action_log.action_template
-                    WHERE action_log.action_log_id = ? AND (actions.action_target = '${ACTION_TARGETS.TEAM}' OR action_log.system_id = ?)`;
-          params = [req.query.log_id, req.user.system_id];
-          break;
-        case ROLES.COACH:
-          getSubmissionQuery = `SELECT action_log.files, action_log.project, action_log.system_id, actions.action_id, actions.action_target
-                    FROM action_log
-                    JOIN actions ON actions.action_id = action_log.action_template
-                    JOIN project_coaches ON project_coaches.project_id = action_log.project
-                    WHERE action_log.action_log_id = ? AND project_coaches.coach_id = ?`;
-          params = [req.query.log_id, req.user.system_id];
-          break;
-        case ROLES.ADMIN:
-          getSubmissionQuery = `SELECT action_log.files, action_log.project, action_log.system_id, actions.action_id, actions.action_target
-                    FROM action_log
-                    JOIN actions ON actions.action_id = action_log.action_template
-                    WHERE action_log.action_log_id = ?`;
-          params = [req.query.log_id];
-          break;
-        default:
-          const error = new Error("Unknown Role");
-          error.statusCode = 401;
-          return next(error);
-      }
-
-      const { files, project, action_target, system_id, action_id } =
-        (await db.query(getSubmissionQuery, params))[0] || {};
-
-      let fileList = [];
-      if (files) {
-        fileList = files.split(",");
-      }
-
-      if (
-        fileList.includes(req.query.file) &&
-        project &&
-        action_target &&
-        system_id &&
-        action_id
-      ) {
-        return res.sendFile(
-          path.join(
-            __dirname,
-            `../project_docs/${project}/${action_target}/${action_id}/${system_id}/${req.query.file}`,
-          ),
-        );
-      }
-      const error = new Error(
-        "File not found or you are unauthorized to view file",
-      );
-      error.statusCode = 404;
-      return next(error);
     },
   );
